Rename APR tooltip component
